requirements.txt
psycopg2-binary
aiohttp
asyncio
discord.py
requests
asgiref
strsimpy